Formulario de puntos

+++ app/Http/Controllers/Recreovia/PuntosController.php
@@ -27,14 +27,14 @@ class PuntosController extends Controller {
 	{
 		$perPage = config('app.page_size');
 		$elementos = Punto::with('localidad', 'upz')
+							->whereNull('deleted_at')
 							->orderBy('Cod_IDRD', 'ASC')
 							->paginate($perPage);
 
 		$lista = [
 			'titulo' => 'Puntos',
 	        'elementos' => $elementos,
-	        'localidades' => Localidad::all(),
-	        'upz' => Upz::all()
+	        'status' => session('status')
 		];
 
 		$datos = [
@@ -43,6 +43,47 @@ class PuntosController extends Controller {
 		];
 
 		return view('list', $datos);
+	}
+
+	public function crear()
+	{
+		$formulario = [
+			'titulo' => 'Crear ó editar puntos',
+	        'punto' => null,
+	        'localidades' => Localidad::all(),
+	        'upz' => Upz::all(),
+	        'status' => session('status')
+	    ];
+
+	    $datos = [
+			'seccion' => 'Puntos',
+			'formulario' => view('idrd.recreovia.formulario-puntos', $formulario)
+		];
+
+		return view('form', $datos);
+	}
+
+	public function editar(Request $request, $id)
+	{
+		$punto = Punto::with(['jornadas' => function($query)
+		{
+			return $query->whereNull('deleted_at');
+		}])->find($id);
+
+		$formulario = [
+			'titulo' => 'Crear ó editar puntos',
+	        'punto' => $punto,
+	        'localidades' => Localidad::all(),
+	        'upz' => Upz::all(),
+	        'status' => session('status')
+	    ];
+
+	    $datos = [
+			'seccion' => 'Puntos',
+			'formulario' => view('idrd.recreovia.formulario-puntos', $formulario)
+		];
+
+		return view('form', $datos);
 	}
 
 	public function asignarPuntos()
@@ -71,9 +112,11 @@ class PuntosController extends Controller {
 
 	public function obtener(Request $request, $id)
 	{
-		$punto = Punto::with(['jornadas' => function($query){
+		$punto = Punto::with(['jornadas' => function($query)
+		{
 			return $query->whereNull('deleted_at');
 		}])->find($id);
+		
 		return response()->json($punto);
 	}
 
@@ -90,11 +133,23 @@ class PuntosController extends Controller {
 		$punto->Cod_Recreovia = $request['Cod_Recreovia'];
 		$punto->Id_Localidad = $request['Id_Localidad'];
 		$punto->Id_Upz = $request['Id_Upz'];
+		$punto->Latitud = $request['Latitud'];
+		$punto->Longitud = $request['Longitud'];
 		$punto->save();
 
 		$this->sincronizarJornadas($request['Jornadas'], $punto);
 
-        return response()->json(['status' => 'ok']);
+       	return redirect('/puntos/editar/'.$punto['Id_Punto'])->with(['status' => 'success']);
+	}
+
+	public function eliminar(Request $request, $id)
+	{
+		$punto = Punto::where('Id_Punto', $id)
+						->first();
+
+		$punto->delete();
+
+		return redirect('/puntos')->with(['status' => 'success']); 
 	}
 
 	private function sincronizarJornadas($jornadas, $punto)
@@ -102,37 +157,40 @@ class PuntosController extends Controller {
 		$jornadas = json_decode($jornadas);
 		$jornadas_ids = [];
 
-		foreach ($jornadas as $j) 
+		if(is_array($jornadas))
 		{
-			if ($j->Id_Jornada != 0)
-				$jornadas_ids[] = $j->Id_Jornada;
-		}
-		
-		$jornadas_eliminadas = Jornada::where('Id_Punto', $punto['Id_Punto'])
-				->whereNotIn('Id_Jornada', $jornadas_ids)
-				->get();
-
-		foreach ($jornadas_eliminadas as $j) 
-		{
-			$j->delete();
-		}
-
-		foreach ($jornadas as $j) 
-		{
-			if ($j->Id_Jornada != 0)
+			foreach ($jornadas as $j) 
 			{
-				$jornada = Jornada::find($j->Id_Jornada);
-			} else {
-				$jornada = new Jornada;
+				if ($j->Id_Jornada != 0)
+					$jornadas_ids[] = $j->Id_Jornada;
+			}
+			
+			$jornadas_eliminadas = Jornada::where('Id_Punto', $punto['Id_Punto'])
+					->whereNotIn('Id_Jornada', $jornadas_ids)
+					->get();
+
+			foreach ($jornadas_eliminadas as $j) 
+			{
+				$j->delete();
 			}
 
-			$jornada->Id_Punto = $punto['Id_Punto'];
-			$jornada->Jornada = $j->Jornada;
-			$jornada->Dias = $j->Dias;
-			$jornada->Inicio = $j->Inicio;
-			$jornada->Fin = $j->Fin;
+			foreach ($jornadas as $j) 
+			{
+				if ($j->Id_Jornada != 0)
+				{
+					$jornada = Jornada::find($j->Id_Jornada);
+				} else {
+					$jornada = new Jornada;
+				}
 
-			$jornada->save();
+				$jornada->Id_Punto = $punto['Id_Punto'];
+				$jornada->Jornada = $j->Jornada;
+				$jornada->Dias = $j->Dias;
+				$jornada->Inicio = $j->Inicio;
+				$jornada->Fin = $j->Fin;
+
+				$jornada->save();
+			}
 		}
 	}
